requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
python-dotenv==0.19.0
landsatxplore==0.12.0
rasterio==1.2.10